Parse Dates
This uses some of the infrastructure added for Intervals to support extracting
the fields of `DATE 'yyyy-mm-dd'`.

It has a similar philosophical position -- it just extracts the fields into a
struct that has numeric fields for the dates, and does very little verification
-- the only thing that it does is verify that month and day are not 0 and fit
into a u8, since the exact number of days in any given month are variable.