Make ast::ObjectName compare based on ident values, add some display methods
`ObjectName`s are semantically supposed to be the value of each of their items,
irrespective of if they were constructed with or without quotes. In fact, the
quotes are used specifically to avoid constructing objects that have weird
characters (like quotes) in their equality.

This commit still preserves the property that the display of an object will use
the quotes that were provided at construction time, but ideally I think we
would check if any special characters (e.g. dots, quoted capital letters) are present and
display using quotes based on that.

This is spurred by MaterializeInc/materialize#685